+++ .azuredevops/pipelines/build-no-tests.yml
@@ -66,7 +66,7 @@ steps:
   displayName: 'Install dotnet-ef'
   condition: always()
   inputs:
-    script: 'dotnet tool install --global dotnet-ef'
+    script: 'dotnet tool install --version 7.0.13 --global dotnet-ef'
 
 - task: CmdLine@2
   displayName: 'Check dotnet-ef version'
