+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.2.5] - 2023-11-29
+### Fixed
+- Refactored code and minor bug fixes
+- The address property in the userinfo endpoint structure is updated from string to object. [Mock Data Recipient Issue 67](https://github.com/ConsumerDataRight/mock-data-recipient/issues/67)
 
 ## [1.2.4] - 2023-08-22
 ### Fixed
