+++ Source/.dockerignore
@@ -24,3 +24,4 @@
 **/values.dev.yaml
 LICENSE
 README.md
+_temp/
