+++ Source/CDR.DataRecipient.API.Logger/CDR.DataRecipient.API.Logger.csproj
@@ -4,9 +4,9 @@
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-	<Version>1.2.4</Version>
-	<FileVersion>1.2.4</FileVersion>
-	<AssemblyVersion>1.2.4</AssemblyVersion>
+	<Version>1.2.5</Version>
+	<FileVersion>1.2.5</FileVersion>
+	<AssemblyVersion>1.2.5</AssemblyVersion>
   </PropertyGroup>
 
   <ItemGroup>
