+++ Source/CDR.DataRecipient.E2ETests/CDR.DataRecipient.E2ETests.csproj
@@ -3,9 +3,9 @@
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
-	<Version>1.2.4</Version>
-	<FileVersion>1.2.4</FileVersion>
-	<AssemblyVersion>1.2.4</AssemblyVersion>
+	<Version>1.2.5</Version>
+	<FileVersion>1.2.5</FileVersion>
+	<AssemblyVersion>1.2.5</AssemblyVersion>
   </PropertyGroup>
 
   <ItemGroup>
