+++ Source/CDR.DataRecipient.Repository.SQL/CDR.DataRecipient.Repository.SQL.csproj
@@ -2,9 +2,9 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-	<Version>1.2.4</Version>
-	<FileVersion>1.2.4</FileVersion>
-	<AssemblyVersion>1.2.4</AssemblyVersion>
+	<Version>1.2.5</Version>
+	<FileVersion>1.2.5</FileVersion>
+	<AssemblyVersion>1.2.5</AssemblyVersion>
   </PropertyGroup>
 
   <ItemGroup>
