+++ Source/CDR.DataRecipient.SDK/Models/UserInfo.cs
@@ -5,7 +5,7 @@ using System.Runtime.Serialization;
 namespace CDR.DataRecipient.SDK.Models
 {
     [Serializable]
-    public class UserInfo : Dictionary<string, string>
+    public class UserInfo : Dictionary<string, object>
     {
         public UserInfo() : base() { }
