+++ Source/CDR.DataRecipient.Web/CDR.DataRecipient.Web.csproj
@@ -5,9 +5,9 @@
 		<RuntimeIdentifiers>win-x64;linux-x64</RuntimeIdentifiers>
 		<UserSecretsId>e18cd40c-c7d1-4df4-8448-82de536a628c</UserSecretsId>
 		<DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-		<Version>1.2.4</Version>
-		<FileVersion>1.2.4</FileVersion>
-		<AssemblyVersion>1.2.4</AssemblyVersion>
+		<Version>1.2.5</Version>
+		<FileVersion>1.2.5</FileVersion>
+		<AssemblyVersion>1.2.5</AssemblyVersion>
 	</PropertyGroup>
 
 	<ItemGroup>
