+++ Source/CDR.DiscoverDataHolders/CDR.DiscoverDataHolders.csproj
@@ -3,9 +3,9 @@
     <TargetFramework>net6.0</TargetFramework>
     <AzureFunctionsVersion>v4</AzureFunctionsVersion>
     <_FunctionsSkipCleanOutput>true</_FunctionsSkipCleanOutput>
-	<Version>1.2.4</Version>
-	<FileVersion>1.2.4</FileVersion>
-	<AssemblyVersion>1.2.4</AssemblyVersion>
+	<Version>1.2.5</Version>
+	<FileVersion>1.2.5</FileVersion>
+	<AssemblyVersion>1.2.5</AssemblyVersion>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Azure.Storage.Queues" Version="12.12.0" />
