+++ Source/DockerCompose/docker-compose.IntegrationTests.yml
@@ -1,0 +1,53 @@
+# Docker compose file used to execute E2E tests for the Mock Data Recipient.
+# It extends docker-compose.TestsBase.yml.
+
+version: '3.8'
+
+name: mock-data-recipient-integration-tests
+
+services:
+  mock-register:
+    container_name: mock-register-mdr-int
+    extends:
+      file: docker-compose.TestsBase.yml
+      service: mock-register
+
+  mock-data-holder:
+    container_name: mock-data-holder-mdr-int
+    extends:
+      file: docker-compose.TestsBase.yml
+      service: mock-data-holder   
+
+  mock-data-holder-energy:
+    container_name: mock-data-holder-energy-mdr-int
+    extends:
+      file: docker-compose.TestsBase.yml
+      service: mock-data-holder-energy    
+
+  mock-data-recipient:
+    container_name: mock-data-recipient-mdr-int
+    extends:
+      file: docker-compose.TestsBase.yml
+      service: mock-data-recipient
+     
+  mssql:
+    container_name: sql-mdr-int
+    extends:
+      file: docker-compose.TestsBase.yml
+      service:  mssql
+      
+  mock-data-recipient-integration-tests:
+    container_name: mock-data-recipient-integration-tests
+    image: mock-data-recipient-integration-tests
+    build:
+      context: ..
+      dockerfile: Dockerfile.integration-tests
+    environment:
+      - ASPNETCORE_ENVIRONMENT=Release
+    volumes:
+      - "./_temp/mock-data-recipient-integration-tests/tmp:/tmp"   
+      - "./_temp/mock-data-recipient-integration-tests/testresults:/testresults"
+    depends_on:    
+      mock-data-recipient:
+        condition: service_healthy
+
