+++ Source/DockerCompose/docker-compose.TestsBase.yml
@@ -1,4 +1,5 @@
-# Docker compose for build pipeline
+# Docker compose base file for E2E and integration tests.
+# This compose file is extended by docker-compose.E2ETests.yml and docker-compose.IntegrationTests.yml files to compose up and execute E2E or integration tests the Mock Data Recipient.
 
 version: '3.8'
 
@@ -13,16 +14,12 @@ services:
       - "7002:7002"
       - "7006:7006"
     environment:
-      - ASPNETCORE_ENVIRONMENT=Release
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user
-    # volumes:
-    #   - "./_temp/mock-register/tmp:/tmp"      
+      - ASPNETCORE_ENVIRONMENT=Release  
     healthcheck:
       test: ((wget --no-check-certificate --no-verbose --spider https://localhost:7006/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:7002/idp/.well-known/openid-configuration)) || exit 1
       timeout: 15s
-      interval: 15s      
+      interval: 15s
       retries: 20
-      start_period: 5s
     depends_on:
       mssql:
         condition: service_healthy
@@ -36,22 +33,19 @@ services:
       - "8001:8001"
       - "8002:8002"
       - "8005:8005"
+      - "3000:3000"
     environment:
       - ASPNETCORE_ENVIRONMENT=Release
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user
-    # volumes:
-    #   - "./_temp/mock-data-holder/tmp:/tmp"      
+      - CdrAuthServer__AccessTokenExpirySeconds=3600
+      - CdrAuthServer__SupportJarmEncryption=true
     healthcheck:
       test: ((wget --no-check-certificate --no-verbose --spider https://localhost:8005/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:8001/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:3000)) || exit 1      
       timeout: 15s
       interval: 15s
-      retries: 20    
-      start_period: 5s
+      retries: 20 
     depends_on:
-      mssql:
-        condition: service_healthy
       mock-register:
-        condition: service_healthy
+        condition: service_healthy        
 
   mock-data-holder-energy:
     container_name: mock-data-holder-energy
@@ -62,24 +56,18 @@ services:
       - "8101:8101"
       - "8102:8102"
       - "8105:8105"
+      - "3100:3100"
     environment:
       - ASPNETCORE_ENVIRONMENT=Release
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user
-    # volumes:
-    #   - "./_temp/mock-data-holder-energy/tmp:/tmp"      
+      - CdrAuthServer__AccessTokenExpirySeconds=3600
     healthcheck:
       test: ((wget --no-check-certificate --no-verbose --spider https://localhost:8105/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:8101/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:3100)) || exit 1      
       timeout: 15s
       interval: 15s
       retries: 20
-      start_period: 5s
     depends_on:
-      mssql:
-        condition: service_healthy
-      mock-register:
-        condition: service_healthy
       mock-data-holder:
-        condition: service_healthy
+        condition: service_healthy       
 
   mock-data-recipient:
     container_name: mock-data-recipient
@@ -88,44 +76,19 @@ services:
     ports: 
       - "9001:9001"
     build:
-      context: .
+      context: ..
       dockerfile: Dockerfile
     environment:
-      - ASPNETCORE_ENVIRONMENT=Release     
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user
-    # volumes:
-    #   - "./_temp/mock-data-recipient/tmp:/tmp"      
+      - ASPNETCORE_ENVIRONMENT=Release
     healthcheck:
       test: wget --no-check-certificate --no-verbose --spider https://localhost:9001/health || exit 1
       timeout: 15s
       interval: 15s
       retries: 20
-      start_period: 10s
     depends_on:
-      mssql:
-        condition: service_healthy
-      mock-register:
-        condition: service_healthy
-      mock-data-holder:
-        condition: service_healthy    
       mock-data-holder-energy:
         condition: service_healthy
-
-  mock-data-recipient-integration-tests:
-    container_name: mock-data-recipient-integration-tests
-    image: mock-data-recipient-integration-tests
-    build:
-      context: .
-      dockerfile: Dockerfile.integration-tests
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Release
-    volumes:
-      - "./_temp/mock-data-recipient-integration-tests/tmp:/tmp"   
-      - "./_temp/mock-data-recipient-integration-tests/testresults:/testresults"
-    depends_on:    
-      mock-data-recipient:
-        condition: service_healthy
-
+      
   mssql:
     container_name: sql1
     image: 'mcr.microsoft.com/mssql/server:2019-latest'
@@ -136,7 +99,6 @@ services:
       - SA_PASSWORD=Pa{}w0rd2019
     healthcheck:
       test: /opt/mssql-tools/bin/sqlcmd -S . -U sa -P "Pa{}w0rd2019" -Q "SELECT 1" || exit 1 
-      timeout: 15s
+      timeout: 5s
       interval: 15s
-      retries: 20
-      start_period: 5s
+      retries: 20
