--- Source/docker-compose.E2ETests.yml
@@ -1,134 +1,0 @@
-# Docker compose for E2E tests
-
-version: '3.8'
-
-services:
-  mock-register:
-    container_name: mock-register
-    image: mock-register
-    hostname: mock-register
-    ports: 
-      - "7000:7000"
-      - "7001:7001"
-      - "7006:7006"
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Release
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user                  
-    # volumes:
-    #   - "./_temp/mock-register/tmp:/tmp"      
-    healthcheck:
-      test: wget --no-check-certificate --no-verbose --spider https://localhost:7006/health || exit 1
-      timeout: 15s
-      interval: 15s
-      retries: 20
-    depends_on:
-      mssql:
-        condition: service_healthy
-
-  mock-data-holder:
-    container_name: mock-data-holder
-    image: mock-data-holder
-    hostname: mock-data-holder
-    ports: 
-      - "8000:8000"
-      - "8001:8001"
-      - "8002:8002"
-      - "8005:8005"
-      # auth ui
-      - "3000:3000"
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Release
-      # tests need expiry of 3600
-      - CdrAuthServer__AccessTokenExpirySeconds=3600
-      - CdrAuthServer__SupportJarmEncryption=true
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user                  
-    # volumes:
-    #   - "./_temp/mock-data-holder/tmp:/tmp"      
-    healthcheck:
-      test: ((wget --no-check-certificate --no-verbose --spider https://localhost:8005/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:8001/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:3000)) || exit 1      
-      timeout: 15s
-      interval: 15s
-      retries: 20 
-    depends_on:
-      mock-register:
-        condition: service_healthy        
-
-  mock-data-holder-energy:
-    container_name: mock-data-holder-energy
-    image: mock-data-holder-energy
-    hostname: mock-data-holder-energy
-    ports: 
-      - "8100:8100"
-      - "8101:8101"
-      - "8102:8102"
-      - "8105:8105"
-      # auth ui
-      - "3100:3100"
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Release
-      # tests need expiry of 3600
-      - CdrAuthServer__AccessTokenExpirySeconds=3600
-      - CdrAuthServer__SupportJarmEncryption=true
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user                  
-    # volumes:
-    #   - "./_temp/mock-data-holder-energy/tmp:/tmp"      
-    healthcheck:
-      test: ((wget --no-check-certificate --no-verbose --spider https://localhost:8105/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:8101/health) && (wget --no-check-certificate --no-verbose --spider https://localhost:3100)) || exit 1      
-      timeout: 15s
-      interval: 15s
-      retries: 20
-    depends_on:
-      mock-data-holder:
-        condition: service_healthy       
-
-  mock-data-recipient:
-    container_name: mock-data-recipient
-    image: mock-data-recipient
-    hostname: mock-data-recipient
-    ports: 
-      - "9001:9001"
-    build:
-      context: .
-      dockerfile: Dockerfile
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Release
-    # FIXME - MJS - commented out volumes below as volume is mounted as 1001:121 (vsts:docker) in build pipeline and causes issue with chown in dockerfile (appuser:appgroup), ie stops container from starting because of different user                  
-    # volumes:
-    #   - "./_temp/mock-data-recipient/tmp:/tmp"      
-    healthcheck:
-      test: wget --no-check-certificate --no-verbose --spider https://localhost:9001/health || exit 1
-      timeout: 15s
-      interval: 15s
-      retries: 20
-    depends_on:
-      mock-data-holder-energy:
-        condition: service_healthy
-
-  mock-data-recipient-e2e-tests:
-    container_name: mock-data-recipient-e2e-tests
-    image: mock-data-recipient-e2e-tests
-    hostname: mock-data-recipient-e2e-tests
-    build:
-      context: .
-      dockerfile: Dockerfile.e2e-tests
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Release
-    volumes:
-      - "./_temp/mock-data-recipient-e2e-tests/testresults:/testresults"
-    depends_on:
-      mock-data-recipient:
-        condition: service_healthy
-      
-  mssql:
-    container_name: sql1
-    image: 'mcr.microsoft.com/mssql/server:2019-latest'
-    ports: 
-      - "1433:1433"
-    environment:
-      - ACCEPT_EULA=Y
-      - SA_PASSWORD=Pa{}w0rd2019
-    healthcheck:
-      test: /opt/mssql-tools/bin/sqlcmd -S . -U sa -P "Pa{}w0rd2019" -Q "SELECT 1" || exit 1 
-      timeout: 5s
-      interval: 15s
-      retries: 20
